+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -197,6 +197,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Fixed player build when HDRP is in the project but not assigned in the graphic settings.
 - Fixed an issue where ambient probe would be black for the first face of a baked reflection probe
 - VFX: Fixed Missing Reference to Visual Effect Graph Runtime Assembly
+- Fixed an issue where rendering done by users in EndCameraRendering would be executed before the main render loop.
 
 ### Changed
 - Color buffer pyramid is not allocated anymore if neither refraction nor distortion are enabled